+++ test/test_search_items_query.rb
@@ -24,7 +24,7 @@ class TestSearchItemsQuery < Petitest::Test
       keywords: %w(
         dummy_keyword1
         dummy_keyword2
-      )
+      ),
     ).to_hash
     assert do
       actual == expected
@@ -52,7 +52,35 @@ class TestSearchItemsQuery < Petitest::Test
       keywords: %w(
         dummy_keyword1
         dummy_keyword2
-      )
+      ),
+    ).to_hash
+    assert do
+      actual == expected
+    end
+    return_time
+  end
+
+  def test_to_hash_with_item_power_option
+    freeze_time
+    expected = {
+      "AssociateTag" => "dummy_associate_tag",
+      "AWSAccessKeyId" => "dummy_access_key_id",
+      "Keywords" => "dummy_keyword1%2Cdummy_keyword2",
+      "Operation" => "ItemSearch",
+      "Power" => "dummy%20power",
+      "SearchIndex" => "All",
+      "Service" => "AWSECommerceService",
+      "Timestamp" => "2000-01-01T00%3A00%3A00Z",
+      "Version" => "2013-08-01",
+    }
+    actual = ::Rapa::Queries::SearchItemsQuery.new(
+      access_key_id: "dummy_access_key_id",
+      associate_tag: "dummy_associate_tag",
+      keywords: %w(
+        dummy_keyword1
+        dummy_keyword2
+      ),
+      power: "dummy power",
     ).to_hash
     assert do
       actual == expected